+++ .travis/test.sh
@@ -6,7 +6,11 @@ set -x
 python -m opem test
 pip install -r dev-requirements.txt
 python version_check.py
-python -m bandit -r opem -s B322
-python -m vulture --min-confidence 80 --exclude=opem,build,.eggs --sort-by-size .
+if [ "$TRAVIS_PYTHON_VERSION" = '3.6' ]
+then
+	python -m bandit -r opem -s B322
+	python -m vulture --min-confidence 80 --exclude=opem,build,.eggs --sort-by-size .
+	python -m pydocstyle
+fi
 python -m pytest opem/Test --cov=opem --cov-report=term
 python -m cProfile -s cumtime opem/Profile.py