+++ opem/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
-'''
-OPEM Init
-'''
+"""OPEM modules."""
+
 import opem.Static.Amphlett
 import opem.Static.Larminie_Dicks
 import opem.Static.Chamberline_Kim
@@ -11,3 +10,4 @@ import opem.Script
 import opem.Dynamic.Padulles1
 import opem.Dynamic.Padulles2
 import opem.Dynamic.Padulles_Hauer
+import opem.Dynamic.Padulles_Amphlett